moved currency autoparse to separate file
